+++ test/unit/fuzz-tools.test.js
@@ -266,6 +266,7 @@ import * as fibrosisV275 from '../../lib/fibrosis-v275.js';
 import * as nutritionV276 from '../../lib/nutrition-v276.js';
 import * as renalV277 from '../../lib/renal-v277.js';
 import * as pedsSepsisV278 from '../../lib/peds-sepsis-v278.js';
+import * as rccPrognosisV279 from '../../lib/rcc-prognosis-v279.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -476,6 +477,7 @@ const MODULES = {
   'nutrition-v276.js': nutritionV276,
   'renal-v277.js': renalV277,
   'peds-sepsis-v278.js': pedsSepsisV278,
+  'rcc-prognosis-v279.js': rccPrognosisV279,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];